fix: 修復 PIXI.js 和 framer-motion 依賴問題
- 將 PixiGame.tsx 簡化為 Canvas 版本
- 移除所有 framer-motion 導入和使用
- 使用 CSS 動畫和 transition 替代 motion 組件
- 修復構建錯誤，確保 Vercel 部署成功
