+++ components/PixiGame.tsx
@@ -1,5 +1,4 @@
 import { useEffect, useRef, useState, useCallback } from 'react';
-import * as PIXI from 'pixi.js';
 
 interface PixiGameProps {
   width?: number;
@@ -14,7 +13,7 @@ interface PixiGameProps {
       color?: number;
     }>;
   };
-  onComplete?: (score: number, total: number) => void;
+  onComplete?: (score?: number, total?: number) => void;
 }
 
 export default function PixiGame({
@@ -24,231 +23,131 @@ export default function PixiGame({
   gameData = { items: [] },
   onComplete
 }: PixiGameProps) {
-  const pixiContainer = useRef<HTMLDivElement>(null);
-  const app = useRef<PIXI.Application | null>(null);
-  const gameObjects = useRef<Map<string, PIXI.Container>>(new Map());
+  const canvasRef = useRef<HTMLCanvasElement>(null);
   const [score, setScore] = useState(0);
   const [gameCompleted, setGameCompleted] = useState(false);
 
-  // 動畫遊戲對象
-  const animateGameObjects = useCallback(() => {
-    gameObjects.current.forEach((obj) => {
-      // 簡單的浮動動畫
-      obj.y += Math.sin(Date.now() * 0.001 + obj.x * 0.01) * 0.5;
-    });
-  }, []);
-
-  // 檢查遊戲完成
-  const checkGameCompletion = useCallback(() => {
-    // 模擬遊戲完成條件
-    const totalItems = gameData.items?.length || 0;
-    const currentScore = Math.floor(Math.random() * totalItems) + 1;
-    
-    setScore(currentScore);
-    
-    if (currentScore >= totalItems) {
-      setGameCompleted(true);
-      onComplete?.(currentScore, totalItems);
-    }
-  }, [gameData.items, onComplete]);
-
-  // 拖動開始
-  const onDragStart = useCallback((event: any) => {
-    const obj = event.currentTarget;
-    obj.alpha = 0.5;
-    obj.data = event.data;
-    obj.dragging = true;
-  }, []);
-
-  // 拖動結束
-  const onDragEnd = useCallback((event: any) => {
-    const obj = event.currentTarget;
-    obj.alpha = 1;
-    obj.dragging = false;
-    obj.data = null;
-    
-    // 檢查遊戲完成
-    checkGameCompletion();
-  }, [checkGameCompletion]);
-
-  // 拖動移動
-  const onDragMove = useCallback((event: any) => {
-    const obj = event.currentTarget;
-    if (obj.dragging) {
-      const newPosition = obj.data.getLocalPosition(obj.parent);
-      obj.x = newPosition.x;
-      obj.y = newPosition.y;
-    }
-  }, []);
-
-  // 更新遊戲元素
+  // 簡化的遊戲邏輯
   const updateGame = useCallback(() => {
-    if (!app.current) return;
-
-    // 清除現有遊戲對象
-    gameObjects.current.forEach((obj) => {
-      app.current?.stage.removeChild(obj);
-    });
-    gameObjects.current.clear();
-
-    // 添加新的遊戲對象
-    if (gameData.items && gameData.items.length > 0) {
-      gameData.items.forEach((item) => {
-        // 創建遊戲對象
-        const container = new PIXI.Container();
-        container.x = item.x || Math.random() * (width - 100);
-        container.y = item.y || Math.random() * (height - 50);
-        container.interactive = true;
-        (container as any).buttonMode = true;
-
-        // 創建背景
-        const background = new PIXI.Graphics();
-        background.beginFill(item.color || 0x3498db);
-        background.drawRoundedRect(0, 0, 100, 50, 10);
-        background.endFill();
-
-        // 創建文本
-        const text = new PIXI.Text(item.text, {
-          fontFamily: 'Arial',
-          fontSize: 16,
-          fill: 0xffffff,
-          align: 'center'
+    if (!canvasRef.current) return;
+    
+    const canvas = canvasRef.current;
+    const ctx = canvas.getContext('2d');
+    
+    if (ctx) {
+      // 清除畫布
+      ctx.clearRect(0, 0, width, height);
+      
+      // 繪製背景
+      ctx.fillStyle = `#${backgroundColor.toString(16).padStart(6, '0')}`;
+      ctx.fillRect(0, 0, width, height);
+      
+      // 繪製遊戲元素
+      if (gameData.items && gameData.items.length > 0) {
+        gameData.items.forEach((item, index) => {
+          const x = item.x || (index * 120 + 60);
+          const y = item.y || (height / 2);
+          
+          // 繪製圓形
+          ctx.fillStyle = item.color ? `#${item.color.toString(16).padStart(6, '0')}` : '#3498db';
+          ctx.beginPath();
+          ctx.arc(x, y, 30, 0, 2 * Math.PI);
+          ctx.fill();
+          
+          // 繪製文字
+          ctx.fillStyle = '#ffffff';
+          ctx.font = '16px Arial';
+          ctx.textAlign = 'center';
+          ctx.fillText(item.text || `Item ${index + 1}`, x, y + 5);
         });
-        text.x = 50 - text.width / 2;
-        text.y = 25 - text.height / 2;
+      } else {
+        // 默認顯示
+        ctx.fillStyle = '#3498db';
+        ctx.fillRect(50, 50, 100, 100);
+        ctx.fillStyle = '#ffffff';
+        ctx.font = '16px Arial';
+        ctx.textAlign = 'center';
+        ctx.fillText('Canvas Game', 100, 105);
+        ctx.fillText('(PIXI.js removed)', 100, 125);
+      }
+    }
+  }, [width, height, backgroundColor, gameData.items]);
 
-        // 添加到容器
-        container.addChild(background);
-        container.addChild(text);
-
-        // 添加拖動事件
-        container.on('pointerdown', onDragStart)
-                 .on('pointerup', onDragEnd)
-                 .on('pointerupoutside', onDragEnd)
-                 .on('pointermove', onDragMove);
-
-        // 添加到舞台
-        if (app.current && app.current.stage) {
-          app.current.stage.addChild(container);
-          gameObjects.current.set(item.id, container);
+  // 處理點擊事件
+  const handleCanvasClick = useCallback((event: React.MouseEvent<HTMLCanvasElement>) => {
+    const canvas = canvasRef.current;
+    if (!canvas) return;
+    
+    const rect = canvas.getBoundingClientRect();
+    const x = event.clientX - rect.left;
+    const y = event.clientY - rect.top;
+    
+    // 簡單的點擊檢測
+    if (gameData.items && gameData.items.length > 0) {
+      gameData.items.forEach((item, index) => {
+        const itemX = item.x || (index * 120 + 60);
+        const itemY = item.y || (height / 2);
+        
+        const distance = Math.sqrt((x - itemX) ** 2 + (y - itemY) ** 2);
+        if (distance <= 30) {
+          setScore(prev => prev + 1);
+          
+          // 檢查遊戲完成
+          if (score + 1 >= gameData.items!.length) {
+            setGameCompleted(true);
+            onComplete?.(score + 1, gameData.items!.length);
+          }
         }
       });
     }
-  }, [gameData.items, width, height, onDragStart, onDragEnd, onDragMove]);
+  }, [gameData.items, height, score, onComplete]);
 
-  // 初始化遊戲
-  const initGame = useCallback(() => {
-    if (!app.current || !app.current.stage || !app.current.ticker) return;
-
-    // 創建遊戲容器
-    const gameContainer = new PIXI.Container();
-    app.current.stage.addChild(gameContainer);
-
-    // 添加遊戲元素
+  // 重新開始遊戲
+  const handleRestart = useCallback(() => {
+    setScore(0);
+    setGameCompleted(false);
     updateGame();
+  }, [updateGame]);
 
-    // 設置遊戲循環
-    app.current.ticker.add(() => {
-      // 遊戲循環邏輯
-      animateGameObjects();
-    });
-  }, [updateGame, animateGameObjects]);
-
-  // 初始化 PixiJS 應用
+  // 初始化 Canvas
   useEffect(() => {
-    if (!pixiContainer.current) return;
-
-    // 在 effect 开始时复制 ref 值
-    const currentGameObjects = gameObjects.current;
-
-    // 創建 PIXI 應用
-    app.current = new PIXI.Application({
-      width,
-      height,
-      backgroundColor,
-      antialias: true,
-      resolution: window.devicePixelRatio || 1,
-    });
-
-    // 將 PIXI 應用添加到 DOM
-    pixiContainer.current.appendChild(app.current.view as HTMLCanvasElement);
-
-    // 初始化遊戲
-    initGame();
-
-    // 清理函數
-    return () => {
-      if (app.current) {
-        app.current.destroy(true, true);
-        app.current = null;
-      }
-      // 使用在 effect 开始时复制的值
-      if (currentGameObjects) {
-        currentGameObjects.clear();
-      }
-    };
-  }, [width, height, backgroundColor, initGame]);
-
-  // 當遊戲數據變化時更新遊戲
-  useEffect(() => {
-    if (app.current) {
-      updateGame();
-    }
-  }, [gameData, updateGame]);
-
-
-
+    updateGame();
+  }, [updateGame]);
 
   return (
-    <div className="pixi-game-container">
-      <div ref={pixiContainer} className="pixi-canvas-container"></div>
+    <div className="canvas-game-container">
+      <canvas 
+        ref={canvasRef}
+        width={width}
+        height={height}
+        className="border border-gray-300 rounded cursor-pointer"
+        style={{ backgroundColor: `#${backgroundColor.toString(16).padStart(6, '0')}` }}
+        onClick={handleCanvasClick}
+      />
+      <div className="mt-4 text-center">
+        <p>得分: {score}/{gameData.items?.length || 0}</p>
+        <p className="text-gray-600">Canvas 遊戲 (PIXI.js 已移除)</p>
+        <button 
+          className="mt-2 bg-blue-500 hover:bg-blue-600 text-white font-bold py-2 px-4 rounded"
+          onClick={handleRestart}
+        >
+          重新開始
+        </button>
+      </div>
       {gameCompleted && (
-        <div className="game-completed-overlay">
-          <div className="game-completed-message">
-            <h2>遊戲完成！</h2>
-            <p>得分: {score}/{gameData.items?.length || 0}</p>
+        <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center">
+          <div className="bg-white p-6 rounded-lg text-center">
+            <h2 className="text-2xl font-bold mb-4">遊戲完成！</h2>
+            <p className="mb-4">得分: {score}/{gameData.items?.length || 0}</p>
             <button 
               className="bg-blue-500 hover:bg-blue-600 text-white font-bold py-2 px-4 rounded"
-              onClick={() => {
-                setGameCompleted(false);
-                updateGame();
-              }}
+              onClick={handleRestart}
             >
               再玩一次
             </button>
           </div>
         </div>
       )}
-      <style jsx>{`
-        .pixi-game-container {
-          position: relative;
-          width: ${width}px;
-          height: ${height}px;
-          margin: 0 auto;
-        }
-        .pixi-canvas-container {
-          width: 100%;
-          height: 100%;
-        }
-        .game-completed-overlay {
-          position: absolute;
-          top: 0;
-          left: 0;
-          width: 100%;
-          height: 100%;
-          background-color: rgba(0, 0, 0, 0.7);
-          display: flex;
-          justify-content: center;
-          align-items: center;
-          z-index: 10;
-        }
-        .game-completed-message {
-          background-color: white;
-          padding: 2rem;
-          border-radius: 0.5rem;
-          text-align: center;
-        }
-      `}</style>
     </div>
   );
 }
