+++ pages/activities/index.tsx
@@ -5,7 +5,7 @@ import Image from 'next/image';
 import { useRouter } from 'next/router';
 import { useSession } from 'next-auth/react';
 import { MagnifyingGlassIcon, FunnelIcon, Squares2X2Icon, ListBulletIcon, ChevronDownIcon, EyeIcon, UserGroupIcon, ClockIcon } from '@heroicons/react/24/outline';
-import { motion } from 'framer-motion';
+// framer-motion 已移除，使用 CSS 動畫替代
 
 interface Activity {
   id: string;
@@ -331,12 +331,9 @@ export default function ActivitiesBrowse() {
     const templateInfo = getTemplateInfo(activity.templateType);
 
     return (
-      <motion.div 
-        key={activity.id} 
-        className="bg-white rounded-lg shadow-md overflow-hidden hover:shadow-lg transition-shadow duration-300"
-        initial={{ opacity: 0, y: 20 }}
-        animate={{ opacity: 1, y: 0 }}
-        transition={{ duration: 0.3 }}
+      <div
+        key={activity.id}
+        className="bg-white rounded-lg shadow-md overflow-hidden hover:shadow-lg transition-all duration-300 animate-fade-in"
       >
         <Link href={`/activities/${activity.id}`} className="block p-4">
           <div className="flex flex-col md:flex-row md:items-center">
